refactor: Move the assign logic into separate folder
This follows the pattern used for other parts of the move app in that
it separates out the code into a self-contained app.

This should allow this part of the app to be maintained a little easier.
