+++ app/move/app/assign/controllers/base.test.js
@@ -1,5 +1,5 @@
-const presenters = require('../../../../common/presenters')
-const CreateBaseController = require('../../app/new/controllers/base')
+const presenters = require('../../../../../common/presenters')
+const CreateBaseController = require('../../new/controllers/base')
 
 const BaseProto = CreateBaseController.prototype
