+++ app/move/app/assign/fields/move-agreed.js
@@ -1,6 +1,6 @@
 const { cloneDeep } = require('lodash')
 
-const moveAgreed = require('./move-agreed')
+const moveAgreed = require('../../../fields/move-agreed')
 
 const moveAgreedAssign = cloneDeep(moveAgreed)
 moveAgreedAssign.items[1].conditional = 'move_not_agreed_instruction'
