+++ app/move/app/assign/fields/move-not-agreed-instruction.js
@@ -1,0 +1,7 @@
+const moveNotAgreedInstruction = {
+  component: 'raw',
+  html:
+    'Complex cases should be discussed and agreed with the receiving prison.</p><p class="govuk-body">This includes:<p><ul class="govuk-list govuk-list--bullet"><li>Segregated prisoners</li><li>Self harm/prisoners on ACCT</li><li>Mental health issues</li><li>Integrated Drug Treatment System</li></ul>',
+}
+
+module.exports = moveNotAgreedInstruction
