+++ app/move/config/index.js
@@ -1,7 +1,5 @@
-const assignConfig = require('./assign.config')
 const updateConfig = require('./update.config')
 
 module.exports = {
-  assignConfig,
   updateConfig,
 }
