+++ app/move/controllers/index.js
@@ -1,11 +1,9 @@
-const assign = require('./assign')
 const confirmation = require('./confirmation')
 const timeline = require('./timeline')
 const update = require('./update')
 const view = require('./view')
 
 module.exports = {
-  assign,
   confirmation,
   timeline,
   update,
