+++ app/move/fields/index.js
@@ -22,9 +22,7 @@ const hasCourtCase = require('./has-court-case')
 const hasDateTo = require('./has-date-to')
 const lastName = require('./last-name')
 const moveAgreed = require('./move-agreed')
-const moveAgreedAssign = require('./move-agreed-assign')
 const moveAgreedBy = require('./move-agreed-by')
-const moveNotAgreedInstruction = require('./move-not-agreed-instruction')
 const moveType = require('./move-type')
 const people = require('./people')
 const policeNationalComputer = require('./police-national-computer')
@@ -34,7 +32,6 @@ const prisonTransferComments = require('./prison-transfer-comments')
 const prisonTransferType = require('./prison-transfer-type')
 const servingYouthSentence = require('./serving-youth-sentence')
 const shouldSaveCourtHearings = require('./should-save-court-hearings')
-const specialVehicleCheck = require('./special-vehicle-check')
 const timeDue = require('./time-due')
 const toLocation = require('./to-location')
 const toLocationCourtAppearance = require('./to-location-court-appearance')
@@ -115,15 +112,7 @@ const updateFields = {
   police_national_computer: policeNationalComputerUpdate,
 }
 
-const assignFields = {
-  ...cloneDeep(createFields),
-  move_agreed: moveAgreedAssign,
-  move_not_agreed_instruction: moveNotAgreedInstruction,
-  special_vehicle_check: specialVehicleCheck,
-}
-
 module.exports = {
-  assignFields,
   createFields,
   updateFields,
 }
