+++ app/move/index.js
@@ -10,13 +10,14 @@ const wizard = require('../../common/middleware/unique-form-wizard')
 const personEscortRecordApp = require('../person-escort-record')
 const youthRiskAssessmentApp = require('../youth-risk-assessment')
 
+const assignApp = require('./app/assign')
 const cancelApp = require('./app/cancel')
 const newApp = require('./app/new')
 const reviewApp = require('./app/review')
 const unassignApp = require('./app/unassign')
-const { assignConfig, updateConfig } = require('./config')
+const { updateConfig } = require('./config')
 const { confirmation, timeline, view } = require('./controllers')
-const { assignFields, updateFields } = require('./fields')
+const { updateFields } = require('./fields')
 const {
   setMove,
   setMoveWithEvents,
@@ -24,7 +25,7 @@ const {
   setYouthRiskAssessment,
   setAllocation,
 } = require('./middleware')
-const { assign: assignSteps, update: updateSteps } = require('./steps')
+const { update: updateSteps } = require('./steps')
 
 // Define param middleware
 router.param('moveId', setMove)
@@ -61,11 +62,7 @@ moveRouter.use(youthRiskAssessmentApp.mountpath, youthRiskAssessmentApp.router)
 moveRouter.use(reviewApp.mountpath, reviewApp.router)
 moveRouter.use(cancelApp.mountpath, cancelApp.router)
 moveRouter.use(unassignApp.mountpath, unassignApp.router)
-moveRouter.use(
-  '/assign',
-  protectRoute('allocation:person:assign'),
-  wizard(assignSteps, assignFields, assignConfig, 'params.moveId')
-)
+moveRouter.use(assignApp.mountpath, assignApp.router)
 
 updateSteps.forEach(updateJourney => {
   const { key, steps } = updateJourney
