+++ app/move/steps/index.js
@@ -1,7 +1,5 @@
-const assign = require('./assign')
 const update = require('./update')
 
 module.exports = {
-  assign,
   update,
 }
